// src/interfaces/Operator.type.ts
export interface Operator {
    CODIGO: number;
    ATIVO: 'SIM' | 'NAO' | null;
    NOME: string;
    LOGIN: string;
    EMAIL: string;
    SENHA: string;
    EXPIRA_EM: string | null;
    NIVEL: 'ATIVO' | 'RECEP' | 'AMBOS' | 'ADMIN' | null;
    HORARIO: number | null;
    ALTERA_SENHA: 'SIM' | 'NAO' | null;
    CODIGO_ERP: string;
    EDITA_CONTATOS: 'SIM' | 'NAO' | null;
    VISUALIZA_COMPRAS: 'SIM' | 'NAO' | null;
    CADASTRO: 'TOTAL' | 'NULOS' | null;
    ULTIMO_LOGIN_INI: string | null;
    ULTIMO_LOGIN_FIM: string | null;
    DATACAD: string | null;
    CODTELEFONIA: string;
    LOGADO: number | null;
    AGENDA_LIG: 'SIM' | 'NAO' | null;
    LIGA_REPRESENTANTE: 'SIM' | 'NAO' | null;
    BANCO: string;
    FILTRA_DDD: 'SIM' | 'NAO' | null;
    FILTRA_ESTADO: 'SIM' | 'NAO' | null;
    ASTERISK_RAMAL: string | null;
    ASTERISK_USERID: string | null;
    ASTERISK_LOGIN: string | null;
    ASTERISK_SENHA: string | null;
    CODEC: string | null;
    ASSINATURA_EMAIL: string | null;
    LIGA_REPRESENTANTE_DIAS: number | null;
    EMAILOPERADOR: string | null;
    SENHAEMAILOPERADOR: string | null;
    EMAIL_EXIBICAO: string | null;
    NOME_EXIBICAO: string | null;
    data_hora_ultima_mensagem_whats: string | null;
    SETOR: number | null;
    OMNI?: number;
    LIMITE_DIARIO_AGENDAMENTO?: number;
    MANUAL: 'SIM' | 'NAO' | null;
}